refactor: migrate Button component imports to new path and remove old Button file
