+++ app/page.tsx
@@ -10,7 +10,7 @@ import { getStringDate } from "@/components/calendar/utils"
 import React, { Suspense } from "react"
 import Spinner from "@/components/assets/Spinner"
 import { TextAnimate } from "@/components/magicui/TextAnimate"
-import { Button } from "@/components/ui/Button"
+import { Button } from "@/components/button/Button"
 import { ScrollButton } from "@/components/ui/ScrollButton"
 
 export default async function Home() {
