+++ components/IPAScheduler.tsx
@@ -4,7 +4,7 @@ import React from "react"
 import { Card, CardContent, CardTitle } from "@/components/ui/Card"
 import { GiDna2 } from "react-icons/gi"
 import CBCBars from "@/components/assets/CBCBars"
-import { Button } from "@/components/ui/Button"
+import { Button } from "@/components/button/Button"
 
 const request_access_url =
   "https://docs.google.com/forms/d/e/1FAIpQLSdJwbhtMznPjR6QrobBG1u-n3AMLnWdigN-NqFepsC7gbb2tw/viewform"
